Fixes needed for code to compile and execute with -D_GLIBCXX_DEBUG.
git-svn-id: https://sourceforge.llnl.gov/svn/repos/samrai_project/trunk/MASTER@920 f45145c3-882e-4954-aa44-00140844b9ef

+++ tools/restart/RedistributedRestartUtility.C
@@ -762,7 +762,7 @@ void RedistributedRestartUtility::readAndWriteBoxLevelRestartData(
    assert(level_in_dbs.size() == 1 || output_dbs.size() == 1);
 #endif
 
-   const int out_size = output_dbs.size();
+   const int out_size = static_cast<int>(output_dbs.size());
 
    //Create an array of level output databases
    std::vector<boost::shared_ptr<tbox::Database> > level_out_dbs(out_size);
